+++ R/utils.R
@@ -126,6 +126,21 @@ html_squish <- function(html) {
 }
 
 mime_base64encode <- function(raw, linewidth = 76L) {
+  if (is.raw(raw)) {
+    log_debug("Input is already raw.")
+  } else {
+    if (file.exists(raw)) {
+      log_debug("Assuming that input is a file.")
+    } else {
+      log_debug("Assuming that input is not a file.")
+      if (is.character(raw)) {
+        raw <- charToRaw(raw)
+      } else {
+        raw <- as.raw(raw)
+      }
+    }
+  }
+
   base64encode(
     raw,
     linewidth,
@@ -172,4 +187,15 @@ wrap_angle_brackets <- function(x) {
   if (!grepl("^<", x)) x <- paste0("<", x)
   if (!grepl(">$", x)) x <- paste0(x, ">")
   x
+}
+
+
+#' Test if list is nested or flat
+#'
+#' @noRd
+#' @param x A list.
+#' @return A Boolean.
+is.nested <- function(x) {
+  stopifnot(is.list(x))
+  any(sapply(x, function(x) any(class(x) == "list")))
 }